Refactor : 로그인 메서드 위치 이전
