+++ src/main/java/techcourse/myblog/web/MainController.java
@@ -7,10 +7,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import techcourse.myblog.domain.service.ArticleService;
 import techcourse.myblog.dto.UserRequestDto;
 
-import javax.servlet.http.HttpSession;
-
-import static techcourse.myblog.web.SessionManager.USER;
-
 @Controller
 public class MainController {
     private final ArticleService articleService;
@@ -29,20 +25,6 @@ public class MainController {
     @GetMapping("/writing")
     public String moveArticleWritePage() {
         return "article-edit";
-    }
-
-    @GetMapping("/login")
-    public String moveLoginPage(HttpSession httpSession) {
-        if (httpSession.getAttribute(USER) != null) {
-            return "redirect:/";
-        }
-        return "login";
-    }
-
-    @GetMapping("/logout")
-    public String logout(HttpSession httpSession) {
-        httpSession.removeAttribute(USER);
-        return "redirect:/";
     }
 
     @GetMapping("/signup")
